fix(logical-width-1): report Spark narrow widths on both doors
LogicalKey arms in crates/repark-spark/src/type_table.rs answer byte,
short, float and binary; Describe is untouched. Twenty cell-driven pins
in test_logical_width_1.py plus the W-3 reclassifications (nullability
flag rows, to_narrow, to_binary, facade-3 bin goldens, census
D7/D8/D9/D18/D19, io-text probe6) to the recorded Spark answers.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

+++ python/repark/tests/test_nullability_2.py
@@ -104,13 +104,13 @@ _CAST_FLAG_ROWS: list[tuple[str, str, bool, str]] = [
         "ts_to_short_o2",
         "SELECT CAST(TIMESTAMP '2020-01-01 00:00:00' AS SMALLINT) AS v",
         True,
-        "int",
+        "smallint",
     ),
     (
         "ts_to_byte_o2",
         "SELECT CAST(TIMESTAMP '2020-01-01 00:00:00' AS TINYINT) AS v",
         True,
-        "int",
+        "tinyint",
     ),
 ]
 
@@ -403,9 +403,9 @@ def test_cast_nullability_matches_spark() -> None:
 def test_narrow_logical_widths_report_wide_per_logical_width_1() -> None:
     session = _spark_session("true")
     try:
-        assert session.sql("SELECT CAST(1 AS SMALLINT) AS v").dtypes == [("v", "int")]
-        assert session.sql("SELECT CAST(1 AS TINYINT) AS v").dtypes == [("v", "int")]
-        assert session.sql("SELECT CAST(1.5 AS FLOAT) AS v").dtypes == [("v", "double")]
+        assert session.sql("SELECT CAST(1 AS SMALLINT) AS v").dtypes == [("v", "smallint")]
+        assert session.sql("SELECT CAST(1 AS TINYINT) AS v").dtypes == [("v", "tinyint")]
+        assert session.sql("SELECT CAST(1.5 AS FLOAT) AS v").dtypes == [("v", "float")]
     finally:
         session.stop()
